+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+
+### Changed
+
+- Move recaptcha into laravel-recaptchav3 package.
+
 ## [3.0.0] - 2025-11-14
 
 _Stable release based on [3.0.0-rc.1]._
@@ -926,6 +932,7 @@ _Stable release based on [0.1.0-rc.1]._
 
 ## [0.0.0] - 2024-09-12
 
+[Unreleased]: https://https://github.com/internetguru/laravel-common/compare/staging...dev
 [3.0.0]: https://https://github.com/internetguru/laravel-common/compare/v2.17.2...v3.0.0
 [3.0.0-rc.1]: https://github.com/internetguru/laravel-common/releases/tag/v2.17.2
 [2.17.2]: https://https://github.com/internetguru/laravel-common/compare/v2.17.1...v2.17.2